+++ framework-utilities/src/test/java/uk/gov/justice/services/framework/utilities/cdi/DummyCdiBean.java
@@ -1,0 +1,4 @@
+package uk.gov.justice.services.framework.utilities.cdi;
+
+public class DummyCdiBean {
+}